backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
celery==5.5.3
redis==6.4.0
SQLAlchemy==2.0.38
